Mistral-Small-4-119B: relocate text core to models/demos/mistral4/ (dedicated demo)
Per the DeepSeek-class serving decision (tt_transformers' GQA Generator can't host MLA),
move the model-local text modules + tests to a dedicated demo, matching models/demos/deepseek_v3/:
  models/demos/mistral4/tt/mistral4_text.py        (TtMistral4MLA/MoE/DecoderLayer/TextModel)
  models/demos/mistral4/tests/test_m4_*.py + m4_text_reference.py
  models/demos/mistral4/README.md
Shared/generic pieces stay in tt_transformers (dequantize_fp8_state_dict in load_checkpoints,
the reused Pixtral vision tower, ModelArgs, TT_CCL). Imports updated; verified: MLA test passes
from the new path (end-to-end 0.99975), all 9 tests collect.

Created with Claude Code loop.
