+++ models/demos/mistral4/tests/m4_text_reference.py
