+++ models/demos/mistral4/tests/test_m4_vision_tower.py
